Use new evaluator to calc logic
